Groundwater Analysis Through Scatter Chart
Our code utilizes the Modified Normalized Difference Water Index (MNDWI), calculated as (Green-SWIR)/(Green+SWIR), to effectively detect water in urban regions with high vegetation cover. MNDWI has been shown to be more effective than NDWI in built-up urban areas due to the presence of vegetation. We used MNDWI in our case because the study area was located in an urban region with a lot of vegetation.

We chose to use GLEAM because it provides high-quality and reliable estimates of evapotranspiration, which is an essential component in groundwater level analysis. By integrating GLEAM with Sentinel-2 images, we were able to create a powerful tool to monitor and analyze groundwater levels in Nepal. Our approach enables identifying areas with low groundwater levels and predicting potential water stress for better water management decisions. Despite some limitations due to cloud cover affecting Sentinel-2 data, our method is cost-effective, efficient, and scalable to other regions facing similar challenges.
The scatter chart shows numerous interpreted data about the selected geographical location (here, Nepal) to infer the varying groundwater levels over the specified period of time. The collected data was in line with the data we received from an ASU site in the region.